spendmend_adk_app/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spendmend-adk"
version = "0.1.0"
description = "Self-improving agent system for Spendmend development tasks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Rawley Fowler", email = "rawley@spendmend.com"}
]
keywords = ["agent", "llm", "adk", "automation", "self-improvement"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Google ADK
    "google-adk>=0.1.0",

    # Core dependencies
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Async support
    "asyncio>=3.4.3",

    # Database (SQLite async driver)
    "aiosqlite>=0.19.0",

    # HTTP clients
    "httpx>=0.24.0",
    "requests>=2.31.0",

    # Jira integration
    "jira>=3.5.0",

    # GitHub integration
    "PyGithub>=2.1.0",

    # Databricks SQL
    "databricks-sql-connector>=2.9.0",
    "databricks-sdk>=0.40.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "click>=8.1.0",  # For future CLI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
]

[project.scripts]
spendmend-agent = "spendmend_adk.main:main"

[project.urls]
Homepage = "https://github.com/spendmend/spendmend-adk"
Documentation = "https://github.com/spendmend/spendmend-adk/blob/main/README.md"
Repository = "https://github.com/spendmend/spendmend-adk"
Issues = "https://github.com/spendmend/spendmend-adk/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
spendmend_adk = ["**/*.md"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "google.adk.*",
    "google.genai.*",
    "jira.*",
    "github.*",
    "databricks.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=spendmend_adk",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
